# aspnetcore-server/src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Describes CallbackApi, a set of methods that will be called by the WFS, to execute external Actions, Conditions and Rules.",
    "version" : "2.0.0",
    "title" : "Callback Server for Workflow Server API",
    "termsOfService" : "https://workflowengine.io/",
    "contact" : {
      "name" : "OptimaJet",
      "url" : "https://workflowengine.io",
      "email" : "sales@optimajet.com"
    },
    "license" : {
      "name" : "Workflow Engine EULA 1.0",
      "url" : "https://workflowengine.io/agreements/eula/"
    }
  },
  "host" : "localhost:5000",
  "basePath" : "/callbackapi",
  "schemes" : [ "http" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
    "/getactions" : {
      "get" : {
        "description" : "Returns the list of Actions that can be called on this server",
        "operationId" : "GetActions",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "List of actions",
            "schema" : {
              "$ref" : "#/definitions/ListOfStringsOperationResult"
            }
          }
        }
      }
    },
    "/getconditions" : {
      "get" : {
        "description" : "Returns the list of Conditions that can be checked by this server",
        "operationId" : "GetConditions",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "List of conditions",
            "schema" : {
              "$ref" : "#/definitions/ListOfStringsOperationResult"
            }
          }
        }
      }
    },
    "/getrules" : {
      "get" : {
        "description" : "Returns the list of Rules that can be checked by this server",
        "operationId" : "GetRules",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "List of rules",
            "schema" : {
              "$ref" : "#/definitions/ListOfStringsOperationResult"
            }
          }
        }
      }
    },
    "/executeaction" : {
      "post" : {
        "description" : "Executes the naming action on the callback server",
        "operationId" : "ExecuteAction",
        "parameters" : [ {
          "in" : "body",
          "name" : "ExecuteParameters",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ExecuteParameters"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Excecution result with new process parameters",
            "schema" : {
              "$ref" : "#/definitions/ActionExecutionResult"
            }
          }
        }
      }
    },
    "/executecondition" : {
      "post" : {
        "description" : "Executes the naming condition on the callback server",
        "operationId" : "ExecuteCondition",
        "parameters" : [ {
          "in" : "body",
          "name" : "ExecuteParameters",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ExecuteParameters"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Condition result",
            "schema" : {
              "$ref" : "#/definitions/BooleanExecutionResult"
            }
          }
        }
      }
    },
    "/checkrule" : {
      "post" : {
        "description" : "Executes the naming rule on the callback server",
        "operationId" : "CheckRule",
        "parameters" : [ {
          "in" : "body",
          "name" : "ExecuteParameters",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ExecuteParameters"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Rule check result",
            "schema" : {
              "$ref" : "#/definitions/BooleanExecutionResult"
            }
          }
        }
      }
    },
    "/getidentities" : {
      "post" : {
        "description" : "Executes the naming rule on the callback server, returns user ids thats satisfy this rule",
        "operationId" : "GetIdentities",
        "parameters" : [ {
          "in" : "body",
          "name" : "ExecuteParameters",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ExecuteParameters"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Rule get result",
            "schema" : {
              "$ref" : "#/definitions/ListOfStringsOperationResult"
            }
          }
        }
      }
    },
    "/generate" : {
      "post" : {
        "description" : "External scheme generator method",
        "operationId" : "Generate",
        "parameters" : [ {
          "in" : "body",
          "name" : "GenerateParameters",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/GeneratorParameters"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "New generated scheme",
            "schema" : {
              "$ref" : "#/definitions/StringOperationResult"
            }
          }
        }
      }
    },
    "/processstatuschanged" : {
      "post" : {
        "description" : "External event handler for ProcessStatusChanged event",
        "operationId" : "ProcessStatusChanged",
        "parameters" : [ {
          "in" : "body",
          "name" : "ProcessStatusChangedParameters",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ProcessStatusChangedParameters"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OperationResult",
            "schema" : {
              "$ref" : "#/definitions/OperationResult"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "OperationResult" : {
      "type" : "object",
      "required" : [ "success" ],
      "properties" : {
        "success" : {
          "type" : "boolean",
          "description" : "Indicates whether the operation was successful or not"
        },
        "error" : {
          "type" : "string",
          "description" : "The short description of the error"
        },
        "message" : {
          "type" : "string",
          "description" : "The details of the error"
        }
      },
      "example" : {
        "success" : true,
        "error" : "error",
        "message" : "message"
      }
    },
    "ListOfStringsOperationResult" : {
      "allOf" : [ {
        "$ref" : "#/definitions/OperationResult"
      }, {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      } ]
    },
    "StringOperationResult" : {
      "allOf" : [ {
        "$ref" : "#/definitions/OperationResult"
      }, {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "string"
          }
        }
      } ]
    },
    "ActionExecutionResult" : {
      "allOf" : [ {
        "$ref" : "#/definitions/OperationResult"
      }, {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/definitions/ProcessParameters"
          }
        }
      } ]
    },
    "BooleanExecutionResult" : {
      "allOf" : [ {
        "$ref" : "#/definitions/OperationResult"
      }, {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "boolean"
          }
        }
      } ]
    },
    "ExecuteParameters" : {
      "type" : "object",
      "properties" : {
        "Name" : {
          "type" : "string"
        },
        "IdentityId" : {
          "type" : "string"
        },
        "Parameter" : {
          "type" : "string"
        },
        "ProcessInstance" : {
          "$ref" : "#/definitions/ProcessInfo"
        }
      },
      "example" : {
        "IdentityId" : "IdentityId",
        "Parameter" : "Parameter",
        "ProcessInstance" : {
          "PreviousStateForDirect" : "PreviousStateForDirect",
          "ActivityName" : "ActivityName",
          "SchemeId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "Transitions" : [ {
            "TransitionClassifier" : "NotSpecified",
            "ToActivityName" : "ToActivityName",
            "TriggerName" : "TriggerName",
            "ActorIdentityId" : "ActorIdentityId",
            "ExecutorIdentityId" : "ExecutorIdentityId",
            "IsFinalised" : true,
            "ProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "FromStateName" : "FromStateName",
            "TransitionTime" : "2000-01-23T04:56:07.000+00:00",
            "ToStateName" : "ToStateName",
            "FromActivityName" : "FromActivityName"
          }, {
            "TransitionClassifier" : "NotSpecified",
            "ToActivityName" : "ToActivityName",
            "TriggerName" : "TriggerName",
            "ActorIdentityId" : "ActorIdentityId",
            "ExecutorIdentityId" : "ExecutorIdentityId",
            "IsFinalised" : true,
            "ProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "FromStateName" : "FromStateName",
            "TransitionTime" : "2000-01-23T04:56:07.000+00:00",
            "ToStateName" : "ToStateName",
            "FromActivityName" : "FromActivityName"
          } ],
          "StateName" : "StateName",
          "ParentProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "RootProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "PreviousStateForReverse" : "PreviousStateForReverse",
          "History" : [ {
            "Order" : 6,
            "InitialState" : "InitialState",
            "Command" : "Command",
            "DestinationState" : "DestinationState",
            "IdentityId" : "IdentityId",
            "Id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "ProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "TransitionTime" : "2000-01-23T04:56:07.000+00:00",
            "AllowedToEmployeeNames" : "AllowedToEmployeeNames"
          }, {
            "Order" : 6,
            "InitialState" : "InitialState",
            "Command" : "Command",
            "DestinationState" : "DestinationState",
            "IdentityId" : "IdentityId",
            "Id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "ProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "TransitionTime" : "2000-01-23T04:56:07.000+00:00",
            "AllowedToEmployeeNames" : "AllowedToEmployeeNames"
          } ],
          "PreviousActivity" : "PreviousActivity",
          "PreviousState" : "PreviousState",
          "ProcessParameters" : { },
          "PreviousActivityForDirect" : "PreviousActivityForDirect",
          "InstanceStatus" : 0,
          "SchemeCode" : "SchemeCode",
          "Id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "PreviousActivityForReverse" : "PreviousActivityForReverse"
        },
        "Name" : "Name"
      }
    },
    "GeneratorParameters" : {
      "type" : "object",
      "required" : [ "Scheme", "SchemeCode", "SchemeId" ],
      "properties" : {
        "SchemeCode" : {
          "type" : "string"
        },
        "SchemeId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "Parameters" : {
          "type" : "string"
        },
        "Scheme" : {
          "type" : "string"
        }
      },
      "example" : {
        "Parameters" : "Parameters",
        "Scheme" : "Scheme",
        "SchemeId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "SchemeCode" : "SchemeCode"
      }
    },
    "ProcessStatusChangedParameters" : {
      "type" : "object",
      "required" : [ "ProcessId", "SchemeCode" ],
      "properties" : {
        "ProcessId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "SchemeCode" : {
          "type" : "string"
        },
        "ProcessInstance" : {
          "$ref" : "#/definitions/ProcessInfo"
        }
      },
      "example" : {
        "SchemeCode" : "SchemeCode",
        "ProcessInstance" : {
          "PreviousStateForDirect" : "PreviousStateForDirect",
          "ActivityName" : "ActivityName",
          "SchemeId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "Transitions" : [ {
            "TransitionClassifier" : "NotSpecified",
            "ToActivityName" : "ToActivityName",
            "TriggerName" : "TriggerName",
            "ActorIdentityId" : "ActorIdentityId",
            "ExecutorIdentityId" : "ExecutorIdentityId",
            "IsFinalised" : true,
            "ProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "FromStateName" : "FromStateName",
            "TransitionTime" : "2000-01-23T04:56:07.000+00:00",
            "ToStateName" : "ToStateName",
            "FromActivityName" : "FromActivityName"
          }, {
            "TransitionClassifier" : "NotSpecified",
            "ToActivityName" : "ToActivityName",
            "TriggerName" : "TriggerName",
            "ActorIdentityId" : "ActorIdentityId",
            "ExecutorIdentityId" : "ExecutorIdentityId",
            "IsFinalised" : true,
            "ProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "FromStateName" : "FromStateName",
            "TransitionTime" : "2000-01-23T04:56:07.000+00:00",
            "ToStateName" : "ToStateName",
            "FromActivityName" : "FromActivityName"
          } ],
          "StateName" : "StateName",
          "ParentProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "RootProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "PreviousStateForReverse" : "PreviousStateForReverse",
          "History" : [ {
            "Order" : 6,
            "InitialState" : "InitialState",
            "Command" : "Command",
            "DestinationState" : "DestinationState",
            "IdentityId" : "IdentityId",
            "Id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "ProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "TransitionTime" : "2000-01-23T04:56:07.000+00:00",
            "AllowedToEmployeeNames" : "AllowedToEmployeeNames"
          }, {
            "Order" : 6,
            "InitialState" : "InitialState",
            "Command" : "Command",
            "DestinationState" : "DestinationState",
            "IdentityId" : "IdentityId",
            "Id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "ProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "TransitionTime" : "2000-01-23T04:56:07.000+00:00",
            "AllowedToEmployeeNames" : "AllowedToEmployeeNames"
          } ],
          "PreviousActivity" : "PreviousActivity",
          "PreviousState" : "PreviousState",
          "ProcessParameters" : { },
          "PreviousActivityForDirect" : "PreviousActivityForDirect",
          "InstanceStatus" : 0,
          "SchemeCode" : "SchemeCode",
          "Id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "PreviousActivityForReverse" : "PreviousActivityForReverse"
        },
        "ProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
      }
    },
    "ProcessInfo" : {
      "type" : "object",
      "required" : [ "ActivityName", "Id", "InstanceStatus", "SchemeCode" ],
      "properties" : {
        "Id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "StateName" : {
          "type" : "string"
        },
        "ActivityName" : {
          "type" : "string"
        },
        "SchemeId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "SchemeCode" : {
          "type" : "string"
        },
        "PreviousState" : {
          "type" : "string"
        },
        "PreviousStateForDirect" : {
          "type" : "string"
        },
        "PreviousStateForReverse" : {
          "type" : "string"
        },
        "PreviousActivity" : {
          "type" : "string"
        },
        "PreviousActivityForDirect" : {
          "type" : "string"
        },
        "PreviousActivityForReverse" : {
          "type" : "string"
        },
        "ParentProcessId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "RootProcessId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "InstanceStatus" : {
          "type" : "integer"
        },
        "Transitions" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Transition"
          }
        },
        "History" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/HistoryItem"
          }
        },
        "ProcessParameters" : {
          "$ref" : "#/definitions/ProcessParameters"
        }
      },
      "example" : {
        "PreviousStateForDirect" : "PreviousStateForDirect",
        "ActivityName" : "ActivityName",
        "SchemeId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "Transitions" : [ {
          "TransitionClassifier" : "NotSpecified",
          "ToActivityName" : "ToActivityName",
          "TriggerName" : "TriggerName",
          "ActorIdentityId" : "ActorIdentityId",
          "ExecutorIdentityId" : "ExecutorIdentityId",
          "IsFinalised" : true,
          "ProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "FromStateName" : "FromStateName",
          "TransitionTime" : "2000-01-23T04:56:07.000+00:00",
          "ToStateName" : "ToStateName",
          "FromActivityName" : "FromActivityName"
        }, {
          "TransitionClassifier" : "NotSpecified",
          "ToActivityName" : "ToActivityName",
          "TriggerName" : "TriggerName",
          "ActorIdentityId" : "ActorIdentityId",
          "ExecutorIdentityId" : "ExecutorIdentityId",
          "IsFinalised" : true,
          "ProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "FromStateName" : "FromStateName",
          "TransitionTime" : "2000-01-23T04:56:07.000+00:00",
          "ToStateName" : "ToStateName",
          "FromActivityName" : "FromActivityName"
        } ],
        "StateName" : "StateName",
        "ParentProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "RootProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "PreviousStateForReverse" : "PreviousStateForReverse",
        "History" : [ {
          "Order" : 6,
          "InitialState" : "InitialState",
          "Command" : "Command",
          "DestinationState" : "DestinationState",
          "IdentityId" : "IdentityId",
          "Id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "ProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "TransitionTime" : "2000-01-23T04:56:07.000+00:00",
          "AllowedToEmployeeNames" : "AllowedToEmployeeNames"
        }, {
          "Order" : 6,
          "InitialState" : "InitialState",
          "Command" : "Command",
          "DestinationState" : "DestinationState",
          "IdentityId" : "IdentityId",
          "Id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "ProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "TransitionTime" : "2000-01-23T04:56:07.000+00:00",
          "AllowedToEmployeeNames" : "AllowedToEmployeeNames"
        } ],
        "PreviousActivity" : "PreviousActivity",
        "PreviousState" : "PreviousState",
        "ProcessParameters" : { },
        "PreviousActivityForDirect" : "PreviousActivityForDirect",
        "InstanceStatus" : 0,
        "SchemeCode" : "SchemeCode",
        "Id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "PreviousActivityForReverse" : "PreviousActivityForReverse"
      }
    },
    "Transition" : {
      "type" : "object",
      "required" : [ "FromActivityName", "ProcessId", "ToActivityName", "TransitionClassifier", "TransitionTime" ],
      "properties" : {
        "ProcessId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "ActorIdentityId" : {
          "type" : "string"
        },
        "ExecutorIdentityId" : {
          "type" : "string"
        },
        "FromActivityName" : {
          "type" : "string"
        },
        "FromStateName" : {
          "type" : "string"
        },
        "IsFinalised" : {
          "type" : "boolean"
        },
        "ToActivityName" : {
          "type" : "string"
        },
        "ToStateName" : {
          "type" : "string"
        },
        "TransitionClassifier" : {
          "type" : "string",
          "enum" : [ "NotSpecified", "Direct", "Reverse" ]
        },
        "TransitionTime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "TriggerName" : {
          "type" : "string"
        }
      },
      "example" : {
        "TransitionClassifier" : "NotSpecified",
        "ToActivityName" : "ToActivityName",
        "TriggerName" : "TriggerName",
        "ActorIdentityId" : "ActorIdentityId",
        "ExecutorIdentityId" : "ExecutorIdentityId",
        "IsFinalised" : true,
        "ProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "FromStateName" : "FromStateName",
        "TransitionTime" : "2000-01-23T04:56:07.000+00:00",
        "ToStateName" : "ToStateName",
        "FromActivityName" : "FromActivityName"
      }
    },
    "HistoryItem" : {
      "type" : "object",
      "required" : [ "DestinationState", "Id", "InitialState", "Order", "ProcessId" ],
      "properties" : {
        "Id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "ProcessId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "IdentityId" : {
          "type" : "string"
        },
        "AllowedToEmployeeNames" : {
          "type" : "string"
        },
        "TransitionTime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "Order" : {
          "type" : "integer"
        },
        "InitialState" : {
          "type" : "string"
        },
        "DestinationState" : {
          "type" : "string"
        },
        "Command" : {
          "type" : "string"
        }
      },
      "example" : {
        "Order" : 6,
        "InitialState" : "InitialState",
        "Command" : "Command",
        "DestinationState" : "DestinationState",
        "IdentityId" : "IdentityId",
        "Id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "ProcessId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "TransitionTime" : "2000-01-23T04:56:07.000+00:00",
        "AllowedToEmployeeNames" : "AllowedToEmployeeNames"
      }
    },
    "ProcessParameters" : {
      "type" : "object"
    }
  }
}